feat: split web files
